core: Dispatch volume changes in protocols
Relates to #1522
